Move HidraProjectFileMode into separate file refs #219
