+++ pyrs/core/reduction_manager.py
@@ -7,10 +7,10 @@ from pyrs.core import instrument_geometry
 from pyrs.core import mask_util
 from pyrs.core import reduce_hb2b_mtd
 from pyrs.core import reduce_hb2b_pyrs
-from pyrs.projectfile import HidraConstants
+from pyrs.projectfile import HidraConstants, HidraProjectFileMode
 from pyrs.utilities import calibration_file_io
 from pyrs.utilities import checkdatatypes
-from pyrs.utilities.rs_project_file import HidraProjectFile, HidraProjectFileMode
+from pyrs.utilities.rs_project_file import HidraProjectFile
 
 
 class HB2BReductionManager(object):
