+++ pyrs/projectfile/__init__.py
@@ -1,4 +1,6 @@
 # flake8: noqa
+# pylint: disable=undefined-variable
 from .constants import HidraConstants
+from .file_mode import HidraProjectFileMode
 
-__all__ = ['HidraConstants']
+__all__ = constants.__all__ + file_mode.__all__
