+++ scripts/preparetest/convert_strain_stress_data.py
@@ -2,7 +2,8 @@
 # System (library) test for classes and methods that will be used to calculate strain and stress
 import os
 import sys
-from pyrs.utilities.rs_project_file import HidraProjectFile, HidraProjectFileMode
+from pyrs.projectfile import HidraProjectFileMode
+from pyrs.utilities.rs_project_file import HidraProjectFile
 
 
 def contain_strain_stress_main():
