+++ tests/data/convert_xray_data.py
@@ -7,10 +7,9 @@ Convert the synchrotron raw data to test
 
 Note: most of the methods to parse HZB data are copied from script pyrscalibration.py
 """
-from pyrs.projectfile import HidraConstants
-from pyrs.utilities.rs_project_file import HidraProjectFile, HidraProjectFileMode
+from pyrs.projectfile import HidraConstants, HidraProjectFileMode
+from pyrs.utilities.rs_project_file import HidraProjectFile
 from pyrs.core.instrument_geometry import HidraSetup
-import numpy
 from skimage import io
 from PIL import Image
 import numpy as np
@@ -96,8 +95,8 @@ def main():
 
     # add sample log data & sub runs
     # for log_name in ['sub-run', '2Theta']
-    project_file.append_experiment_log(HidraConstants.SUB_RUNS, numpy.array([1]))
-    project_file.append_experiment_log(HidraConstants.TWO_THETA, numpy.array([35.]))
+    project_file.append_experiment_log(HidraConstants.SUB_RUNS, np.array([1]))
+    project_file.append_experiment_log(HidraConstants.TWO_THETA, np.array([35.]))
 
     # add instrument information
     instrument_setup = generate_xray_instrument()
